fix(quickload): two clippy errors the pre-push gate compiles for
`type_complexity` on the new `map_place_names` table, which is a genuine
readability complaint: the key and the value each mean something, so they get
names -- `MapKey` is the three bytes both param tables carry and `NamedPlace` is
a resolved id plus the table it came from.

`unnecessary_cast` in `product_continue`: `found.slot` is already `usize`, so the
cast said nothing and hid that the types had converged.

The gate that caught these compiles every pushed commit rather than the working
tree, which is why neither showed up locally until the push.

+++ crates/er-quickload/src/experiments/continue_load/product_continue.rs
@@ -362,7 +362,7 @@ fn picker_detail_for_configured_save(slot: i32) -> ConfiguredSaveTruth {
     }
     match slots.iter().find(|info| info.slot as i32 == slot) {
         Some(found) => ConfiguredSaveTruth {
-            loadable: Some((path, found.slot as usize)),
+            loadable: Some((path, found.slot)),
             detail: format!(
                 "{name} slot {slot} holds {} at level {}, so the save itself is fine -- this mod failed to start the load.",
                 found.name, found.level
